backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.13.4
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.2.0
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
openpyxl==3.0.10
packaging==24.0
parsel==1.9.1
pluggy==1.5.0
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
pytest==8.1.1
pytest-django==4.8.0
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.34.0
requests==2.31.0
requests-file==2.0.0
rpds-py==0.18.0
Scrapy==2.11.1
scrapy-djangoitem==1.1.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tldextract==5.1.2
tomli==2.0.1
Twisted==24.3.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.3
